Add files via upload
This task has been performed by Marcus Bowen on the 2nd of March. It is free for use by anyone for practice. No one should submit it for marking in a school set up. 
I was planning to implement a simple signup &login for student portal then implement a database to store the emails and names which will allow for registered student to login. That will achieve all the requirements in the order i.e the simple app with two activities and a fragment for each..then  parse data using JSOn package to a database. I will use mongoDB.